bench_fastapi/requirements.txt
fastapi[all]
asyncpg
pyjwt
gunicorn
pyjwt
python-dotenv[cli]
databases[postgresql]